+++ lib/features/user/services/user_follow_genre_service.dart
@@ -1,4 +1,4 @@
-// user_follow_genre_service.dart
+// lib/features/user/services/user_follow_genre_service.dart
 
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:sway/features/genre/models/genre_model.dart';
@@ -7,27 +7,52 @@ import 'package:sway/features/user/models/user_model.dart' as AppUser;
 import 'package:sway/features/user/services/user_service.dart';
 
 class UserFollowGenreService {
-  final _supabase = Supabase.instance.client;
-  final int userId = 3; // L'ID de l'utilisateur actuel
+  final SupabaseClient _supabase = Supabase.instance.client;
+  final UserService _userService = UserService();
 
+  /// Récupère l'ID de l'utilisateur actuellement connecté
+  Future<int?> _getCurrentUserId() async {
+    final currentUser = await _userService.getCurrentUser();
+    return currentUser?.id;
+  }
+
+  /// Vérifie si l'utilisateur suit un genree spécifique
   Future<bool> isFollowingGenre(int genreId) async {
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
     final response = await _supabase
         .from('user_follow_genre')
         .select()
         .eq('user_id', userId)
-        .eq('genre_id', genreId);
+        .eq('genre_id', genreId)
+        .single();
 
     return response.isNotEmpty;
   }
 
+  /// Suit un genree
   Future<void> followGenre(int genreId) async {
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
     await _supabase.from('user_follow_genre').insert({
       'user_id': userId,
       'genre_id': genreId,
     });
   }
 
+  /// Ne suit plus un genree
   Future<void> unfollowGenre(int genreId) async {
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
     await _supabase
         .from('user_follow_genre')
         .delete()
@@ -35,15 +60,18 @@ class UserFollowGenreService {
         .eq('genre_id', genreId);
   }
 
+  /// Récupère le nombre de followers d'un genree
   Future<int> getGenreFollowersCount(int genreId) async {
     final response = await _supabase
         .from('user_follow_genre')
         .select('user_id')
-        .eq('genre_id', genreId);
+        .eq('genre_id', genreId)
+        .single();
 
     return response.length;
   }
 
+  /// Récupère les genrees suivis par un utilisateur spécifique
   Future<List<Genre>> getFollowedGenresByUserId(int userId) async {
     final response = await _supabase
         .from('user_follow_genre')
@@ -60,15 +88,16 @@ class UserFollowGenreService {
         .toList();
   }
 
-  Future<List<AppUser.User>> getUsersFollowingGenre(int genreId) async {
+  /// Récupère les utilisateurs qui suivent un genree spécifique
+  Future<List<AppUser.User>> getFollowersForGenre(int genreId) async {
     final response = await _supabase
         .from('user_follow_genre')
         .select('user_id')
         .eq('genre_id', genreId);
 
-    final List<int> userIds =
+    final List<int> followerIds =
         response.map((item) => item['user_id'] as int).toList();
 
-    return await UserService().getUsersByIds(userIds);
+    return await _userService.getUsersByIds(followerIds);
   }
 }
